commiting policy updates and missing partial _

+++ app/controllers/wikis_controller.rb
@@ -6,14 +6,17 @@ class WikisController < ApplicationController
 
   def show
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
     @wiki = Wiki.new(wiki_params)
+    authorize @wiki
     if @wiki.save
        flash.now[:notice] = "Your wiki was saved."
        redirect_to @wiki
@@ -25,12 +28,13 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
     @wiki.assign_attributes(wiki_params)
-      authorize @wiki
     if @wiki.save
        flash.now[:notice] = "Your wiki was updated."
        redirect_to @wiki
@@ -42,12 +46,14 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
-    if @wiki.destroy
-      flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
-      redirect_to wikis_path
-    else
-      flash.now[:alert] = "There was an error deleting the wiki."
-      render :show
+    if current_user == @wiki.user
+      if @wiki.destroy
+        flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
+        redirect_to wikis_path
+      else
+        flash.now[:alert] = "There was an error deleting the wiki."
+        render :show
+      end
     end
   end
